% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vh_get.R
\name{vh_route}
\alias{vh_route}
\title{Get route from Valhalla API}
\usage{
vh_route(
  from,
  to,
  costing = "pedestrian",
  directions_options = list(units = "km"),
  url = "http://localhost:8002",
  ...
)
}
\arguments{
\item{from}{A numeric vector of length 2, specifying the longitude and latitude of the starting point.}

\item{to}{A numeric vector of length 2, specifying the longitude and latitude of the destination.}

\item{costing}{A string specifying the costing model to use for route optimization. Default is "auto".}

\item{directions_options}{A named list.
By default this is list(units = "km") specifying the units to use.}

\item{url}{A string specifying the URL of the Valhalla API. Default is "http://localhost:8002/optimized_route".}

\item{...}{Additional parameters to pass to the Valhalla API.}
}
\value{
An sf object representing the route.
}
\description{
Get route from Valhalla API
}
\examples{
if (FALSE) {
  andorra_la_vella = c(1.5218, 42.5075)
  pas_de_la_casa = c(1.7333, 42.5425)
  r1 = vh_get(from = andorra_la_vella, to = pas_de_la_casa)
}
}
