% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vh_get.R
\name{vh_get}
\alias{vh_get}
\title{Get optimized route from Valhalla API}
\usage{
vh_get(
  from,
  to,
  costing = "auto",
  units = "miles",
  url = "http://localhost:8002/optimized_route",
  ...,
  resource = "route"
)
}
\arguments{
\item{from}{A numeric vector of length 2, specifying the longitude and latitude of the starting point.}

\item{to}{A numeric vector of length 2, specifying the longitude and latitude of the destination.}

\item{costing}{A string specifying the costing model to use for route optimization. Default is "auto".}

\item{units}{A string specifying the units to use for the route distance. Default is "miles".}

\item{url}{A string specifying the URL of the Valhalla API. Default is "http://localhost:8002/optimized_route".}

\item{...}{Additional arguments to pass to httr2::GET.}
}
\value{
An sf object representing the optimized route.
}
\description{
Get optimized route from Valhalla API
}
\examples{
if (FALSE) {

}
}
